refactor: jwtUserService 로직 분리
