+++ src/main/java/gift/user/service/JwtUserService.java
@@ -5,11 +5,9 @@ import gift.user.jwt.JwtService;
 import gift.user.model.UserRepository;
 import gift.user.model.dto.AppUser;
 import gift.user.model.dto.LoginRequest;
-import gift.user.model.dto.SignUpRequest;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import security.SHA256Util;
 
 @Service
 public class JwtUserService {
@@ -19,16 +17,6 @@ public class JwtUserService {
     public JwtUserService(UserRepository userRepository, JwtService jwtService) {
         this.userRepository = userRepository;
         this.jwtService = jwtService;
-    }
-
-    @Transactional
-    public void signUp(SignUpRequest signUpRequest) {
-        String salt = SHA256Util.getSalt();
-        String hashedPassword = SHA256Util.encodePassword(signUpRequest.getPassword(), salt);
-        signUpRequest.setPassword(hashedPassword);
-        AppUser appUser = new AppUser(signUpRequest.getEmail(), signUpRequest.getPassword(), signUpRequest.getRole(),
-                salt);
-        userRepository.save(appUser);
     }
 
     @Transactional(readOnly = true)
